Backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cohere==5.3.3
colorama==0.4.6
dataclasses-json==0.6.4
exceptiongroup==1.2.1
fastapi==0.110.2
fastavro==1.9.4
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
hyperframe==6.0.1
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-cohere==0.1.4
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langsmith==0.1.51
marshmallow==3.21.1
mthrottle==0.0.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.1
packaging==23.2
playwright==1.43.0
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
pyparsing==3.1.2
pypdf==4.2.0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
qdrant-client==1.9.0
requests==2.31.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tenacity==8.2.3
tokenizers==0.19.1
tqdm==4.66.2
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
